Massive overhaul by Eloi George Added comment history

+++ mod/comments/class/Comment_Item.php
@@ -57,7 +57,16 @@ class Comment_Item {
     public $author = null;
 
     // Error encountered when processing object
-    public $_error	      = null;
+    public $_error       = null;
+
+    // Userid of author of comment this comment is a child of
+    public $parent_author_id = 0;
+
+    // UnregName of author of comment this comment is a child of
+    public $parent_anon_name = null;
+
+    // If set, this comment is locked to this thread.  It cannot be moved or deleted.
+    public $protected      = 0;
 
     public function __construct($id=null)
     {
@@ -86,7 +95,6 @@ class Comment_Item {
         }
     }
 
-
     public function setId($id)
     {
 	$this->id = (int)$id;
@@ -107,7 +115,6 @@ class Comment_Item {
 	$this->parent = (int)$parent;
     }
 
-
     public function setSubject($subject)
     {
 	$this->subject = strip_tags(trim($subject));
@@ -123,9 +130,9 @@ class Comment_Item {
     {
 	if ($format) {
             $entry = PHPWS_Text::parseOutput($this->entry, true, true);
-	} else {
-	    $entry =  $this->entry;
-	}
+        } else {
+            $entry =  $this->entry;
+        }
 
         if ($quoted) {
             return sprintf('[quote="%s"]%s[/quote]', $this->getAuthorName(), trim($entry));
@@ -261,10 +268,16 @@ class Comment_Item {
 	return $this->_error;
     }
 
-    public function getTpl($allow_anon, $can_post=true)
+    public function getTpl(&$thread)
     {
         $author = $this->getAuthor();
-	$author_info = $author->getTpl();
+
+        // If anonymous users are allowed to post or
+        // the current user is logged in and not banned from posting
+        $can_post = $thread->canComment();
+
+        $is_moderator = empty($thread->phpwsbb_topic) || PHPWSBB_Forum::isModerator($author->user_id, $thread->phpwsbb_topic->fid);
+        $author_info = $author->getTpl($is_moderator);
 
         if (!$this->author_id && $this->anon_name) {
             $author_info['AUTHOR_NAME'] = & $this->anon_name;
@@ -284,19 +297,31 @@ class Comment_Item {
         if ($can_post) {
             $template['QUOTE_LINK']  = $this->quoteLink();
             $template['REPLY_LINK']  = $this->replyLink();
-            if (!isset($_SESSION['Users_Reported_Comments'][$this->id])) {
-                $template['REPORT_LINK'] = $this->reportLink();
-            } else {
-                $template['REPORT_LINK'] = dgettext('comments', 'Reported!');
+            if (Current_User::isLogged()) {
+                if (empty($_SESSION['Users_Reported_Comments'][$this->id]))
+                    $template['REPORT_LINK'] = $this->reportLink();
+                else
+                    $template['REPORT_LINK'] = dgettext('comments', 'Reported!');
+            }
+            if ($thread->userCan() || ($this->author_id > 0 && $this->author_id == Current_User::getId()))
+                $template['EDIT_LINK']       = $this->editLink();
+            if (!$this->protected && $thread->userCan('delete_comments'))
+                $template['DELETE_LINK']     = $this->deleteLink();
+            // If phpwsbb is installed && we have moderation & forking privileges
+            if (isset($GLOBALS['Modules']['phpwsbb']) && $thread->userCan('fork_messages', 'phpwsbb')) {
+                $vars = array();
+                $vars['op'] = 'split_topic';
+                $vars['split_point'] = $this->create_time;
+                $vars['topic'] = $this->thread_id;
+                $str = dgettext('comments', 'Split topic from this position');
+                $str = '<span>' . dgettext('comments', 'Split Topic') . '</span>';
+                $title = dgettext('comments', 'Split topic from this position');
+                $template['FORK_THIS'] = PHPWS_Text::secureLink($str, 'phpwsbb', $vars, null, $title, 'comment_fork_link');
             }
         }
-
-        if ($can_post) {
-            $template['EDIT_LINK']    = $this->editLink();
-        }
-        $template['DELETE_LINK']  = $this->deleteLink();
-        $template['VIEW_LINK']    = $this->viewLink();
-        $template['PUNISH_LINK']     = $this->punishUserLink();
+        $template['VIEW_LINK']       = $this->viewLink();
+        if ($thread->userCan('punish_users'))
+            $template['PUNISH_LINK']     = $this->punishUserLink();
 
         if ($this->parent) {
             $template['RESPONSE_LABEL']  = dgettext('comments', 'In response to');
@@ -323,12 +348,18 @@ class Comment_Item {
         }
 
         $template['ANCHOR'] = sprintf('<a name="cm_%s"></a>', $this->id);
+        $str = dgettext('comments', 'Back to Top');
+        $template['TO_TOP'] = sprintf('<a href="%s#comments"><img src="./images/mod/comments/back_to_top.png" title="%s" border="0" height="16" width="16"> %s</a>', $_SERVER['REQUEST_URI'],$str, $str);
+        $template['COMMENT_ID'] = $this->id;
 
-	if (Current_User::allow('comments')) {
-	    $template['IP_ADDRESS'] = $this->getIp();
-	}
-	$template = array_merge($author_info, $template);
-	return $template;
+        if ($thread->userCan('delete_comments')) {
+            $template['IP_ADDRESS_LABEL'] = dgettext('comments', 'IP');
+            $template['IP_ADDRESS'] = $this->getIp();
+            if (!$this->protected)
+                $template['SELECT_THIS'] = '<input type="checkbox" name="cm_id[]" value="'.$this->id.'" title="'.dgettext('comments', 'Select this Comment').'">';
+        }
+        $template = array_merge($author_info, $template);
+        return $template;
     }
 
     public function save($stamp_update=true)
@@ -345,11 +376,11 @@ class Comment_Item {
 	    $this->stampCreateTime();
 	}
 
-	if (!$this->id) {
-	    $this->stampIP();
-	    $this->stampAuthor();
-	    $increase_count = TRUE;
-	} else {
+        if (empty($this->id)) {
+    	    $this->stampIP();
+    	    $this->stampAuthor();
+    	    $increase_count = TRUE;
+    	} else {
             if ($stamp_update) {
                 $this->stampEditor();
             }
@@ -369,82 +400,96 @@ class Comment_Item {
         $thread = new Comment_Thread($this->thread_id);
         $thread->increaseCount();
         $thread->postLastUser($this->author_id);
-        return $thread->save();
+        $result = $thread->save();
+        if (!PHPWS_Error::logIfError($result)) {
+            // Update any associated phpwsbb topic's lastpost information
+            if (!empty($thread->phpwsbb_topic)) {
+                $thread->phpwsbb_topic->update_topic();
+                $thread->phpwsbb_topic->commit();
+            }
+            // Start subscription?
+            $user = Comments::getCommentUser(Current_User::getId());
+            if (!$thread->monitored && $user->monitordefault)
+                Comment_User::subscribe(Current_User::getId(), $thread->id);
+            // Send notices to all subscribed users
+            Comments::sendUpdateNotice($thread, $this);
+        }
+        return $result;
     }
 
     public function editLink()
     {
-	if (Current_User::allow('comments') ||
-	    ($this->author_id > 0 && $this->author_id == Current_User::getId())
-	    ) {
-	    $vars['uop']   = 'post_comment';
-	    $vars['thread_id']	   = $this->thread_id;
-	    $vars['cm_id']	   = $this->id;
-	    return PHPWS_Text::moduleLink(dgettext('comments', 'Edit'), 'comments', $vars);
-	} else {
-	    return null;
-	}
+        $vars['uop']   = 'post_comment';
+        $vars['cm_id'] = $this->id;
+        $vars['thread_id'] = $this->thread_id;
+        $str = '<span>' . dgettext('comments', 'Edit') . '</span>';
+        $title = dgettext('comments', 'Edit this comment');
+        return PHPWS_Text::secureLink($str, 'comments', $vars, null, $title, 'comment_edit_link');
     }
 
     public function deleteLink()
     {
-	if (Current_User::allow('comments', 'delete_comments')) {
-	    $vars['QUESTION'] = dgettext('comments', 'Are you sure you want to delete this comment?');
-	    $vars['ADDRESS'] = 'index.php?module=comments&amp;cm_id=' . $this->id . '&amp;aop=delete_comment&amp;authkey='
-		. Current_User::getAuthKey();
-	    $vars['LINK'] = dgettext('comments', 'Delete');
-	    return Layout::getJavascript('confirm', $vars);
-	} else {
-	    return null;
-	}
-
+        $vars['QUESTION'] = dgettext('comments', 'Are you sure you want to delete this comment?');
+        $vars['ADDRESS'] = 'index.php?module=comments&amp;cm_id=' . $this->id . '&amp;aop=delete_comment&amp;authkey='
+            . Current_User::getAuthKey();
+        $vars['LINK'] = '<span>' . dgettext('comments', 'Delete') . '</span>';
+        $vars['CLASS'] = 'comment_delete_link';
+        $vars['TITLE'] = dgettext('comments', 'Delete this comment');
+        return Layout::getJavascript('confirm', $vars);
     }
 
     public function clearReportLink()
     {
         return PHPWS_Text::secureLink(dgettext('comments', 'Clear'), 'comments',
-                                      array('aop'=>'clear_report', 'cm_id'=>$this->id));
+                                      array('aop'=>'clear_report', 'cm_id'=>$this->id),
+                                      NULL, dgettext('comments', 'Clear this report'));
     }
 
     public function punishUserLink($graphic=false)
     {
-        if (Current_User::allow('comments', 'punish_users')) {
-            $vars['address'] = PHPWS_Text::linkAddress('comments', array('aop'=>'punish_user',
-                                                                         'cm_id'=>$this->id), true);
+        $vars['address'] = PHPWS_Text::linkAddress('comments', array('aop'=>'punish_user',
+                                                                     'cm_id'=>$this->id, 'authkey'=>Current_User::getAuthKey()), true);
+        $vars['link_title'] = dgettext('comments', 'Punish this user');
+        $vars['class'] = 'comment_punish_link';
             if ($graphic) {
                 $vars['label'] = sprintf('<img src="images/mod/comments/noentry.png" width="20" height="20" title="%s" alt="%s"/>',
                                          dgettext('comments', 'Punish poster'), dgettext('comments', 'Punish icon'));
             } else {
-                $vars['label'] = dgettext('comments', 'Punish');
+                $vars['label'] = dgettext('comments', 'Punish this user');
             }
-            $vars['width'] = 240;
-            $vars['height'] = 180;
-            return javascript('open_window', $vars);
-        } else {
-            return null;
-        }
+
+        $vars['width'] = 240;
+        $vars['height'] = 180;
+        return javascript('open_window', $vars);
     }
 
     public function quoteLink()
     {
-	$vars['uop']   = 'post_comment';
-	$vars['thread_id']     = $this->thread_id;
-	$vars['cm_parent']     = $this->id;
-	return PHPWS_Text::moduleLink(dgettext('comments', 'Quote'), 'comments', $vars);
+        $vars['uop']   = 'post_comment';
+        $vars['thread_id'] = $this->thread_id;
+        $vars['cm_parent'] = $this->id;
+        $vars['type']      = 'quote';
+        $str = '<span>' . dgettext('comments', 'Quote') . '</span>';
+        $title = dgettext('comments', 'Quote this comment');
+        return PHPWS_Text::moduleLink($str, 'comments', $vars, null, $title, 'comment_quote_link');
     }
 
     public function replyLink()
     {
-	$vars['uop']   = 'post_comment';
-	$vars['thread_id']     = $this->thread_id;
-	return PHPWS_Text::moduleLink(dgettext('comments', 'Reply'), 'comments', $vars);
+        $vars['uop']   = 'post_comment';
+        $vars['thread_id'] = $this->thread_id;
+        $vars['cm_parent'] = $this->id;
+        $str = '<span>' . dgettext('comments', 'Reply') . '</span>';
+        $title = dgettext('comments', 'Reply to this comment');
+        return PHPWS_Text::secureLink($str, 'comments', $vars, null, $title, 'comment_reply_link');
     }
 
     public function reportLink()
     {
-        return sprintf('<a href="#" onclick="report(%s, this); return false">%s</a>',
-                       $this->id,
-                       dgettext('comments', 'Report'));
+        $str = '<span>' . dgettext('comments', 'Report') . '</span>';
+        $title = dgettext('comments', 'Report this comment to an administrator');
+        return sprintf('<a href="#" class="%s" onclick="report(%s, this); return false" title="%s">%s</a>',
+                       'comment_report_link', $this->id, $title, $str);
     }
 
     public function viewLink()
@@ -460,20 +505,52 @@ class Comment_Item {
      */
     public function delete($reduce_count=true)
     {
+        // Protected comments cannot be deleted
+        if ($this->protected)
+            return false;
+
         // physical removal
-        $thread = new Comment_Thread($this->thread_id);
         $db = new PHPWS_DB('comments_items');
         $db->addWhere('id', $this->id);
         if (PHPWS_Error::logIfError($db->delete())) {
             return false;
         }
 
-        // clear replies to this comment
+        // orphan replies to this comment
         $this->clearChildren();
 
+        $thread = new Comment_Thread($this->thread_id);
         // decrease thread count
         if ($reduce_count) {
             $thread->decreaseCount();
+
+            // Update parent thread's lastposter info
+            // Update any associated phpwsbb topic's lastpost information
+            if (!empty($thread->phpwsbb_topic)) {
+                $thread->phpwsbb_topic->update_topic();
+                $thread->phpwsbb_topic->commit();
+                $thread->last_poster = $thread->phpwsbb_topic->lastpost_author;
+            }
+            else {
+                $db = new PHPWS_DB('comments_items');
+                $db->addColumn('comments_items.author_id');
+                $db->addColumn('comments_items.anon_name');
+                $db->addColumn('users.display_name');
+                $db->addWhere('users.id', 'comments_items.author_id');
+                $db->addWhere('thread_id', $this->thread_id);
+                $db->addWhere('comments_items.approved', 1);
+                $db->addOrder('create_time desc');
+                $result = $db->select('row');
+                if (PHPWS_Error::logIfError($result))
+                    return;
+                if (empty($result))
+                    $thread->last_poster = dgettext('comments', 'None');
+                elseif ($row['display_name'])
+                    $thread->last_poster = $row['display_name'];
+                else
+                    $thread->last_poster = trim($row['anon_name'] . ' '.COMMENT_ANONYMOUS_TAG);
+            }
+
             $thread->save();
         }
         return true;
@@ -497,10 +574,23 @@ class Comment_Item {
 
     public function responseAuthor()
     {
-        $comment = new Comment_Item($this->parent);
-	$vars['uop']   = 'view_comment';
-	$vars['cm_id'] = $comment->id;
-	return PHPWS_Text::moduleLink($comment->getAuthorName(), 'comments', $vars);
+        if (!empty($this->parent_author_id)) {
+            $author = Comments::getCommentUser($this->parent_author_id);
+            $name = $author->display_name;
+        }
+        else if (!empty($this->parent_anon_name))
+            $name = $this->parent_anon_name.' '.COMMENT_ANONYMOUS_TAG;
+        else { // backward compatibility - cache data for future use
+            $comment = new Comment_Item($this->parent);
+            $name = $comment->getAuthorName();
+            $this->parent_author_id = $comment->author_id;
+            $this->parent_anon_name = $comment->anon_name;
+            $db = new PHPWS_DB('comments_items');
+            PHPWS_Error::logIfError($db->saveObject($this));
+        }
+        $vars['uop']   = 'view_comment';
+        $vars['cm_id'] = $this->parent;
+        return PHPWS_Text::moduleLink($name, 'comments', $vars);
     }
 
     public function reportTags()
@@ -517,6 +607,28 @@ class Comment_Item {
         $links[] = $this->deleteLink();
         $links[] = $this->punishUserLink();
         $tpl['ACTION']  = implode(' | ', $links);
+        return $tpl;
+    }
+
+    public function historyTags()
+    {
+        $tpl['SUBJECT'] = $this->viewLink();
+
+        $tpl['ENTRY']   = sprintf('<span class="pointer" onmouseout="quick_view(\'#cm%s\'); return false" onmouseover="quick_view(\'#cm%s\'); return false">%s</span>',
+                                  $this->id, $this->id,
+                                  substr($this->entry, 0, 50));
+        $tpl['FULL'] = sprintf('<div class="full-view" id="cm%s">%s</div>', $this->id, $this->getEntry());
+
+	$tpl['POSTED_ON']	   = dgettext('comments', 'Posted on');
+	$tpl['CREATE_TIME']   = $this->getCreateTime();
+
+        if (Current_User::allow('comments')) {
+            $tpl['CHECK'] = sprintf('<input type="checkbox" name="cm_id[]" value="%s" />', $this->id);
+            $links[] = $this->clearReportLink();
+            $links[] = $this->deleteLink();
+            $links[] = $this->punishUserLink();
+            $tpl['ACTION']  = implode(' | ', $links);
+        }
         return $tpl;
     }
 
@@ -537,8 +649,8 @@ class Comment_Item {
                            dgettext('comments', 'Approval icon'));
 
         $remove = sprintf('<img src="images/mod/comments/cancel.png" width="20" height="20" title="%s" alt="%s" />',
-                           dgettext('comments', 'Remove this comment'),
-                           dgettext('comments', 'Removal icon'));
+                          dgettext('comments', 'Remove this comment'),
+                          dgettext('comments', 'Removal icon'));
 
 
         $links[] = PHPWS_Text::secureLink($approve, 'comments', array('aop'=>'approve',
